dto restructure
