+++ src/main/java/com/archons/springwildparkapi/controller/v1/AuthenticationControllerV1.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.archons.springwildparkapi.dto.AuthenticationRequest;
-import com.archons.springwildparkapi.dto.AuthenticationResponse;
-import com.archons.springwildparkapi.dto.RegisterAccountRequest;
+import com.archons.springwildparkapi.dto.reesponses.AuthenticationResponse;
+import com.archons.springwildparkapi.dto.requests.AuthenticationRequest;
+import com.archons.springwildparkapi.dto.requests.RegisterAccountRequest;
 import com.archons.springwildparkapi.exceptions.AccountNotFoundException;
 import com.archons.springwildparkapi.exceptions.DuplicateEntityException;
 import com.archons.springwildparkapi.service.AccountService;
