+++ src/main/java/com/archons/springwildparkapi/dto/reesponses/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.reesponses;
 
 import com.archons.springwildparkapi.model.AccountEntity;
