+++ src/main/java/com/archons/springwildparkapi/dto/requests/AddBookingRequest.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.requests;
 
 public class AddBookingRequest {
     private int vehicleId;
