+++ src/main/java/com/archons/springwildparkapi/dto/requests/AddVehicleRequest.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.requests;
 
 import com.archons.springwildparkapi.model.FourWheelVehicleType;
 import com.archons.springwildparkapi.model.VehicleType;
