+++ src/main/java/com/archons/springwildparkapi/dto/requests/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.requests;
 
 public class AuthenticationRequest {
     private String email;
