+++ src/main/java/com/archons/springwildparkapi/dto/requests/UpdateOrganizationRequest.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.requests;
 
 import com.archons.springwildparkapi.model.AccountEntity;
 import com.archons.springwildparkapi.model.OrganizationType;
