+++ src/main/java/com/archons/springwildparkapi/dto/requests/UpdateParkingAreaRequest.java
@@ -1,4 +1,4 @@
-package com.archons.springwildparkapi.dto;
+package com.archons.springwildparkapi.dto.requests;
 
 public class UpdateParkingAreaRequest {
     private int slots;
